refactored carousel
