+++ frontend/src/components/Navbar/index.tsx
@@ -1,0 +1,5 @@
+import React from "react";
+
+const Navbar = () => {
+  
+}
